[Move #254] SpeakTimer.tsx 생성
